fix: absolute paths for retrieve json
