+++ src/utils/coverage.ts
@@ -17,6 +17,7 @@ import {
 } from '@salesforce/apex-node';
 import { Successes, Failures, CodeCoverage } from '@salesforce/source-deploy-retrieve';
 import { ensureArray } from '@salesforce/kit';
+import { StandardColors } from '@salesforce/sf-plugins-core';
 
 type SuccessOrFailure = Successes & Failures;
 
@@ -95,4 +96,33 @@ export const transformCoverageToApexCoverage = (mdCoverage: CodeCoverage[]): Ape
     return ac;
   });
   return { done: true, totalSize: apexCoverage.length, records: apexCoverage };
+};
+
+export const coverageOutput = (
+  cov: CodeCoverage
+): Pick<CodeCoverage, 'name' | 'numLocations'> & { lineNotCovered: string } => {
+  const numLocationsNum = parseInt(cov.numLocations, 10);
+  const numLocationsNotCovered: number = parseInt(cov.numLocationsNotCovered, 10);
+  const color = numLocationsNotCovered > 0 ? StandardColors.error : StandardColors.success;
+
+  let pctCovered = 100;
+  const coverageDecimal: number = parseFloat(((numLocationsNum - numLocationsNotCovered) / numLocationsNum).toFixed(2));
+  if (numLocationsNum > 0) {
+    pctCovered = coverageDecimal * 100;
+  }
+  // cov.numLocations = color(`${pctCovered}%`);
+  const base = {
+    name: cov.name,
+    numLocations: color(`${pctCovered}%`),
+  };
+
+  if (!cov.locationsNotCovered) {
+    return { ...base, lineNotCovered: '' };
+  }
+  const locations = ensureArray(cov.locationsNotCovered);
+
+  return {
+    ...base,
+    lineNotCovered: locations.map((location) => location.line).join(','),
+  };
 };
